refactor: component -> components 폴더 오탈자 수정
