+++ src/app/router.tsx
@@ -2,9 +2,9 @@ import { createBrowserRouter, Outlet } from 'react-router-dom';
 import ANALYSIS_ROUTES from './routes/ANALYSIS_ROUTES';
 import HOME_ROUTES from './routes/HOME_ROUTES';
 import MANAGEMENT_ROUTES from './routes/MANAGEMENT_ROUTES';
-import AnalysisLayout from '../component/layouts/containerLayout/AnalysisLayout';
-import HomeLayout from '../component/layouts/containerLayout/HomeLayout';
-import ManagementLayout from '../component/layouts/containerLayout/ManagementLayout';
+import AnalysisLayout from '../components/layouts/containerLayout/AnalysisLayout';
+import HomeLayout from '../components/layouts/containerLayout/HomeLayout';
+import ManagementLayout from '../components/layouts/containerLayout/ManagementLayout';
 
 const router = createBrowserRouter(
     [
