+++ src/containers/home/common/TopBar/buttons/PersonalButton.tsx
@@ -2,7 +2,7 @@ import { useNavigate } from 'react-router-dom';
 import LoginIcon from '@mui/icons-material/Login';
 import PersonOutlineOutlinedIcon from '@mui/icons-material/PersonOutlineOutlined';
 import PATH from '../../../../../domain/constants/path';
-import RoundedButton from '../../../../../component/common/RoundedButton';
+import RoundedButton from '../../../../../components/common/RoundedButton';
 
 function PersonalButton() {
   const navigate = useNavigate();
